mediacapture-extensions/MediaStreamTrack-getConstraints.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/mediacapture-image/resources/imagecapture-helpers.js"></script>
<script>

const constraints = {
  backgroundBlur: true,
};

// These tests verify that MediaStreamTrack.getConstraints() exists and that,
// returns the constraints passed beforehand with applyConstraints.
function makePromiseTest(name, constraint) {
  image_capture_test(async (t) => {
    const stream = await navigator.mediaDevices.getUserMedia({video: true});
    const [videoTrack] = stream.getVideoTracks();

    const constraintsIn = {advanced: [constraint]};
    await videoTrack.applyConstraints(constraintsIn);
    assert_object_equals(videoTrack.getConstraints(), constraintsIn,
                         'constraints');

    // Clear constraints by sending an empty constraint set.
    await videoTrack.applyConstraints({});
    assert_object_equals(videoTrack.getConstraints(), {}, 'constraints');
  }, name);
};

// Send each line of |constraints| in turn and then the whole dictionary.
for (key in constraints) {
  const one_constraint = {[key]: constraints[key]};
  makePromiseTest(`MediaStreamTrack.getConstraints(), key: ${key}`,
                  one_constraint);
}

makePromiseTest('MediaStreamTrack.getConstraints(), complete', constraints);
</script>
